Add Arrthorizer::Rails::Configuration
This class will load the config YAML file and convert it to a runtime
"configuration" - that is, create the proper objects (like permissions).

This will be done in a config.to_prepare to ensure that Roles and such
are only loaded at that time, so that Rails' development mode reloading
does not break. This is, for now, a Rails-specific class, which might be
completely or partially extracted to a more generic class later on.

Refs #17